AddrSequence: bug fix: fail to update refer count when the address is not aligned to page size
the size of cluster end - address may not aligned (inclulded less than page size case) to page
size sometimes, which will cause the delta value is wrong is cluster, cause later we can NOT
update the refer count of the address. Now for these cases we will put the address into a
new cluster.