Refactor class ETPMigrate for better multi-thread support
We separate the hot/cold page selection algorithm and anti-thrahsing
from the class EPTMigrate in this path, these changes make EPTMigrate
to be a pure exectution or implementation class, don't combine with
any higher logic and policy, which allow it to work better in
mulit-thread situation.

The new structure migrate_parameter is used for saving all necessary
migration parameter, the new page selection algorithm will set them
in later patches. It also included an indicator to enable/disable
migration, anti-thrahsing will use this in later patches.