Apply limitation for cold page and change some behavior of balanced state in calc_migrate_count()
This patch also changed the behavior for balanced state:
Set both demote_limit_kb and promote_limit_kb to 0 When either avail_cold_page_kb
or avail_hot_page_kb is 0.

We should do some small step adjustment in balance state, so set demote and
promote count to 0 here.

Let's have a look whether this is a good choice, in POV of profile tool and sys-refs.